Added second AbbreviateLine method
Also fixed the newline-adding of AbbreviateLine() to behave better.

This was for testing purposes.

Testing overview:
- Timed with call to time
- Input A = /usr/share/dict/words (many lines, one word per line)
- Input B = man c++ (many lines, many more words per line on average)

Using some very crude tests, the second method seems a little bit faster
for input A, and the first one seems quite a bit faster for input B.

I'll probably revert the second method.