Reimplemented A8e() to use regex findall()
I don't know if this is more or less efficient than the previous way of
doing this, but it sure hurts my brain less.  (I suspect there may be
smarter ways to handle the final string substitutions in A8e(), but
it'll do for now.)

Since it no longer disassembles & reassembles the string, A8e() now
respects whitespace better:  fixes #1.